More event handling.

+++ source/blender/editors/space_armory/ArmoryWrapper.cpp
@@ -72,6 +72,51 @@ namespace {
 	std::map<std::string, bool> shaderChanges;
 	std::map<std::string, std::string> shaderFileNames;
 
+    void update();
+    void keyDown(Kore::KeyCode code, wchar_t character);
+    void keyUp(Kore::KeyCode code, wchar_t character);
+    void mouseMove(int window, int x, int y, int mx, int my);
+    void mouseDown(int window, int button, int x, int y);
+    void mouseUp(int window, int button, int x, int y);
+    
+    void krom_init(const v8::FunctionCallbackInfo<v8::Value>& args) {
+//        HandleScope scope(args.GetIsolate());
+//        Local<Value> arg = args[0];
+//        String::Utf8Value title(arg);
+//        int width = args[1]->ToInt32()->Value();
+//        int height = args[2]->ToInt32()->Value();
+//        int samplesPerPixel = args[3]->ToInt32()->Value();
+//        
+//        Kore::WindowOptions options;
+//        options.title = *title;
+//        options.width = width;
+//        options.height = height;
+//        options.x = 100;
+//        options.y = 100;
+//        options.targetDisplay = 0;
+//        options.mode = Kore::WindowModeWindow;
+//        options.rendererOptions.depthBufferBits = 16;
+//        options.rendererOptions.stencilBufferBits = 8;
+//        options.rendererOptions.textureFormat = 0;
+//        options.rendererOptions.antialiasing = samplesPerPixel;
+//        Kore::System::initWindow(options);
+//        
+//        Kore::Graphics::setRenderState(Kore::DepthTest, false);
+//        //Mixer::init();
+//        //Audio::init();
+//        Kore::Random::init(Kore::System::time() * 1000);
+//        
+//        Kore::System::setCallback(update);
+//        
+//        Kore::Keyboard::the()->KeyDown = keyDown;
+//        Kore::Keyboard::the()->KeyUp = keyUp;
+//        Kore::Mouse::the()->Move = mouseMove;
+//        Kore::Mouse::the()->Press = mouseDown;
+//        Kore::Mouse::the()->Release = mouseUp;
+        
+        //Mixer::play(music);
+    }
+    
 	void LogCallback(const v8::FunctionCallbackInfo<v8::Value>& args) {
 		if (args.Length() < 1) return;
 		HandleScope scope(args.GetIsolate());
@@ -1213,6 +1258,7 @@ namespace {
 		HandleScope handle_scope(isolate);
 
 		Local<ObjectTemplate> krom = ObjectTemplate::New(isolate);
+        krom->Set(String::NewFromUtf8(isolate, "init"), FunctionTemplate::New(isolate, krom_init));
 		krom->Set(String::NewFromUtf8(isolate, "log"), FunctionTemplate::New(isolate, LogCallback));
 		krom->Set(String::NewFromUtf8(isolate, "clear"), FunctionTemplate::New(isolate, graphics_clear));
 		krom->Set(String::NewFromUtf8(isolate, "setCallback"), FunctionTemplate::New(isolate, krom_set_callback));
@@ -1640,8 +1686,22 @@ void armoryNew() {
 }
 
 bool first = true;
+bool good = false;
+
+void filesLocationChanged() {
+	good = false;
+	Kore::setFilesLocation(armory_url);
+}
+
 void armoryShow(int x, int y, int w, int h) {	
 	
+    std::ifstream f(armory_url);
+	if (!f.good()) {
+		good = false;
+		return;
+	}
+	good = true;
+
 	if (first) {
 		first = false;
 
@@ -1649,7 +1709,6 @@ void armoryShow(int x, int y, int w, int h) {
 		// shadersdir = armory_url + "-resources";
 		// kromjs = assetsdir + "/krom.js";
 		
-		Kore::setFilesLocation(armory_url);
 		Kore::System::setName("Krom");
 		Kore::System::setup();
 		Kore::WindowOptions options;
@@ -1696,11 +1755,13 @@ void armoryShow(int x, int y, int w, int h) {
 }
 
 void armoryExit() {
+	if (!good) return;
 	armory_started = false;
 	startKrom("armory.Data.deleteAll();");
 }
 
 void armoryDraw() {
+	if (!good) return;
 	update();
 	//Kore::System::callback();
 	//Kore::System::handleMessages();
@@ -1711,6 +1772,7 @@ bool armoryStarted() {
 }
 
 void armoryUpdatePosition(int x, int y, int w, int h) {
+	if (!good) return;
 	// Kore::System::setWindowWidth(0, w);
 	// Kore::System::setWindowHeight(0, h);
 }
@@ -1720,19 +1782,23 @@ void armoryFree() {
 }
 
 void armoryCallJS() {
+	if (!good) return;
 	startKrom(armory_jssource);
 }
 
 void armoryMouseMove(int x, int y) {
+	if (!good) return;
 	Kore::Mouse::the()->_move(0, x, y);
 }
 
 void armoryMousePress(int button, int x, int y) {
+	if (!good) return;
 	// window, button
 	Kore::Mouse::the()->_press(0, button, x, y);
 }
 
 void armoryMouseRelease(int button, int x, int y) {
+	if (!good) return;
 	Kore::Mouse::the()->_release(0, button, x, y);
 }
